requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
httpx==0.27.2
redis==5.1.1
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.9
structlog==24.4.0
tenacity==9.0.0
pytest==8.3.3
pytest-asyncio==0.24.0
pandas==2.2.3
openpyxl==3.1.5
